# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PagoPa Cruscotto Backend API",
    "description": "PagoPa Cruscotto Backend API documentation",
    "termsOfService": "",
    "license": {
      "name": "unlicensed",
      "url": ""
    },
    "version": "0.6.3"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/auth-groups": {
      "get": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "getAllAuthGroups",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "descrizione",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthGroupDTO"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "updateAuthGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthGroupDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthGroupDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "createAuthGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthGroupDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthGroupDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth-groups/{idGroup}/associa-funzioni": {
      "post": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "aggiungiAssociazioneFunzione",
        "parameters": [
          {
            "name": "idGroup",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AuthFunctionDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth-functions/{idFunction}/associa-permessi": {
      "post": {
        "tags": [
          "auth-function-resource"
        ],
        "operationId": "aggiungiAssociazioneFunzione_1",
        "parameters": [
          {
            "name": "idFunction",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AuthPermissionDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/auth-groups/{id}": {
      "get": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "getAuthGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthGroupDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "deleteAuthGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth-groups/{idGroup}/rimuovi-funzione/{funzioneId}": {
      "get": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "rimuoviAssociazioneFunzione",
        "parameters": [
          {
            "name": "idGroup",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "funzioneId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth-groups/detail/{id}": {
      "get": {
        "tags": [
          "auth-group-resource"
        ],
        "operationId": "getAuthGroupWithFunction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthGroupDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth-functions/{idFunction}/rimuovi-permesso/{idPermesso}": {
      "get": {
        "tags": [
          "auth-function-resource"
        ],
        "operationId": "rimuoviAssociazioneFunzione_1",
        "parameters": [
          {
            "name": "idFunction",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idPermesso",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/auth-functions/auth-group/{idGroup}": {
      "get": {
        "tags": [
          "auth-function-resource"
        ],
        "operationId": "getAllAuthFunctions_1",
        "parameters": [
          {
            "name": "idGroup",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthFunctionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthPermissionDTO": {
        "required": [
          "modulo",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "modulo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "AuthFunctionDTO": {
        "required": [
          "descrizione",
          "modulo",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "modulo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "descrizione": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "groupId": {
            "type": "integer",
            "format": "int64"
          },
          "selected": {
            "type": "boolean"
          },
          "authPermissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthPermissionDTO"
            }
          }
        }
      },
      "AuthGroupDTO": {
        "required": [
          "descrizione",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "descrizione": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "livelloVisibilita": {
            "type": "integer",
            "format": "int32"
          },
          "authFunctions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthFunctionDTO"
            }
          }
        }
      },
...
    }
  }
}